+++ python/htpclient/download.py
@@ -7,15 +7,16 @@ from htpclient.initialize import Initialize
 
 class Download:
     @staticmethod
-    def download(url, output):
+    def download(url, output, no_header=False):
         if Initialize.get_os() == 1:
             output = output.replace("/", '\\')
 
         # Check header
-        head = requests.head(url)
-        # not sure if we only should allow 200, but then it's present for sure
-        if head.status_code != 200:
-            return False
+        if not no_header:
+            head = requests.head(url)
+            # not sure if we only should allow 200, but then it's present for sure
+            if head.status_code != 200:
+                return False
 
         with open(output, "wb") as file:
             response = requests.get(url, stream=True)