+++ python/changelog.md
@@ -4,3 +4,5 @@
 * On the logfile output timestamp and log level are now reported
 * Fixed sending error function
 * Fixed bug with killing hashcat process in case the server sends 'stop' or an error occurs
+* Changed running directory for hashcat cracker, changed all paths
+* Extracting is overwriting existing files if present to make sure files are up-to-date